Crate nphysics2d

nphysics

nphysics is a 2 and 3-dimensional physics engine for games and animations. It uses ncollide for collision detection, and nalgebra for vector/matrix math. 2D and 3D implementations both share the same code!

Examples are available in the examples directory. There is also a short (outdated) demonstration video. An on-line version of this documentation is available here. Feel free to ask for help and discuss features on the official user forum.

Why another physics engine?

There are a lot of physics engine out there. However having a physics engine written in Rust is much more fun than writing bindings and has several advantages:

it shows that Rust is suitable for soft real-time applications

it shows how well Rust behaves with highly generic code

it shows that there is no need to write two separate engine for 2D and 3D: genericity wrt the dimension is possible (modulo low level arithmetic specializations for each dimension).

in a not-that-near future, C++ will die of ugliness. Then, people will search for a physics engine and nphysics will be there, proudly exhibiting its Rusty sexyness.

Compilation

You will need the latest release of the Rust compiler and the official package manager: Cargo.

If you want to use the 2D version of nphysics, add the crate named nphysics2d to your dependencies:

[dependencies]
nphysics2d = "0.1.*"

For the 3D version, add the crate named nphysics3d:

[dependencies]
nphysics3d = "0.1.*"

By default, 32-bit floating point numbers are used by the library. If you need more accuracy, use either version of nphysics with the feature f64 enabled. For example:

[dependencies.nphysics2d]
version  = "0.1.*"
features = "f64"

Use make examples to build the demos and execute ./your_favorite_example_here --help to see all the cool stuffs you can do.

Features

static and dynamic rigid bodies

common convex primitives: cone, box, ball, cylinder

concave geometries build from convex primitives (aka. compound geometries)

stable stacking

island based sleeping (objects deactivation)

ray casting

swept sphere based continuous collision detection

ball-in-socket joint

fixed joint

What is missing?

nphysics is a very young library and needs to learn a lot of things to become a grown up. Many missing features are because of missing features on ncollide. Features missing from nphysics itself include:

kinematic bodies

efficient signaling system

more joints, joint limits, joint motors and breakable joints.

soft-bodies (see https://github.com/natal/roft for a draft)

parallel pipeline

GPU-based pipeline

Dependencies

All dependencies are automatically cloned with a recursive clone. The libraries needed to compile the physics engine are:

ncollide: the collision detection library.

nalgebra: the linear algebra library.

The libraries needed to compile the examples are:

kiss3d: the 3d graphics engine.

rust-sfml: the 2D graphics engine. See windows installation instructions.

Modules

aliases
Aliases for complicated parameterized types.

detection
Collision detection and joints.

integration
Position and orientation update of rigid bodies.

math
Compilation flags dependent aliases for mathematical types.

object
Rigid bodies.

resolution
Constraint resolution.

utils
Miscellaneous utilities.

volumetric
Volume and inertia tensor computation.

world
The physics world.