Struct nphysics2d::algebra::Force2

pub struct Force2<N: Real> {
    pub linear: Vector2<N>,
    pub angular: N,
}

A force with a linear and agular (torque) component.

Fields
linear: Vector2<N>

The linear force.

angular: N

The torque.

Methods

impl<N: Real> Force2<N>

pub fn new(linear: Vector2<N>, angular: N) -> Self

Creates a force from its linear and angular components.

pub fn zero() -> Self

A zero force.

pub fn from_slice(data: &[N]) -> Self

Create a force from a slice where the entries 0 and 1 are for the linear part and 2 for the angular part.

pub fn from_vector<S: Storage<N, U3>>(data: &Vector<N, U3, S>) -> Self

Create a force from a vector where the entries 0 and 1 are for the linear part and 2 for the angular part.

pub fn torque(torque: N) -> Self

Create a pure torque.

pub fn torque_from_vector(torque: Vector1<N>) -> Self

Create a pure torque.

pub fn linear(linear: Vector2<N>) -> Self

Create a pure linear force.

pub fn linear_at_point(linear: Vector2<N>, point: &Point2<N>) -> Self

Creates the resultant of a linear force applied at the given point (relative to the center of mass).

pub fn torque_at_point(torque: N, point: &Point2<N>) -> Self

Creates the resultant of a torque applied at the given point (relative to the center of mass).

pub fn torque_from_vector_at_point(
torque: Vector1<N>,
point: &Point2<N>
) -> Self

Creates the resultant of a torque applied at the given point (relative to the center of mass).

Important traits for &'a mut [u8]
impl<'a> Write for &'a mut [u8]impl<'a> Read for &'a [u8]

pub fn as_slice(&self) -> &[N]

This force seen as a slice.

The two first entries contain the linear part and the third entry contais the angular part.

pub fn as_vector(&self) -> &Vector3<N>

This force seen as a vector.

The two first entries contain the linear part and the third entry contais the angular part.

pub fn as_vector_mut(&mut self) -> &mut Vector3<N>

This force seen as a mutable vector.

The two first entries contain the linear part and the third entry contais the angular part.