Struct nphysics2d::algebra::Inertia2

pub struct Inertia2<N: Real> {
    pub linear: N,
    pub angular: N,
}

The inertia of a rigid body grouping both its mass and its angular inertia.

Fields
linear: N

The linear part (mass) of the inertia.

angular: N

The angular inertia.

Methods

impl<N: Real> Inertia2<N>

pub fn new(linear: N, angular: N) -> Self

Creates an inertia from its linear and angular components.

pub fn new_with_angular_matrix(linear: N, angular: Matrix1<N>) -> Self

Creates an inertia from its linear and angular components.

pub fn mass(&self) -> N

Get the mass.

pub fn zero() -> Self

Create a zero inertia.

pub fn angular_matrix(&self) -> &Matrix1<N>

Get the angular inertia tensor.

pub fn to_matrix(&self) -> Matrix3<N>

Convert the inertia into a matrix where the mass is represented as a 2x2 diagonal matrix on the upper-left corner, and the angular part as a 1x1 matrix on the lower-rigth corner.

pub fn transformed(&self, _: &Isometry2<N>) -> Self

Compute the inertia on the given coordinate frame.

pub fn inverse(&self) -> Self

Inverts this inetia matrix.

Sets the angular part to zero if it is not invertible.