Struct nphysics2d::detection::joint::Anchor

pub struct Anchor<N: Scalar, P> {
    pub body: Option<Rc<RefCell<RigidBody<N>>>>,
    pub position: P,
}

One of the two end points of a joint.

Fields
body: Option<Rc<RefCell<RigidBody<N>>>>

The body attached to this anchor.

position: P

The attach position, in local coordinates of the attached body.

Methods

impl<N: Scalar, P> Anchor<N, P>

fn new(body: Option<Rc<RefCell<RigidBody<N>>>>, position: P) -> Anchor<N, P>

Creates a new Anchor at a given position on a body local space.

If body is None, the anchor is concidered to be attached to the ground and position is the attach point in global coordinates.

impl<N: Scalar, P> Anchor<N, P>

fn center_of_mass(&self) -> Point<N>

The center of mass of the body attached to this anchor.

Returns the zero vector if no body is attached.