Struct nphysics2d::detection::joint::JointManager

pub struct JointManager<N: Scalar> {
    // some fields omitted
}

Structure that handles creation and removal of joints.

Methods

impl<N: Scalar> JointManager<N>

fn new() -> JointManager<N>

Creates a new JointManager.

fn joints(&self) -> &HashMap<usize, Constraint<N>, UintTWHash>

Joints handled by this manager.

fn joints_with_body(&self, body: &Rc<RefCell<RigidBody<N>>>) -> Option<&[Constraint<N>]>

List of joints attached to a specific body.

fn add_ball_in_socket(&mut self, joint: Rc<RefCell<BallInSocket<N>>>, activation: &mut ActivationManager<N>)

Add a BallInSocket joint to this manager.

This will force the activation of the two objects attached to the joint.

fn remove_ball_in_socket(&mut self, joint: &Rc<RefCell<BallInSocket<N>>>, activation: &mut ActivationManager<N>)

Removes a BallInSocket joint from this manager.

This will force the activation of the two objects attached to the joint.

fn add_fixed(&mut self, joint: Rc<RefCell<Fixed<N>>>, activation: &mut ActivationManager<N>)

Add a Fixed joint to this manager.

This will force the activation of the two objects attached to the joint.

fn remove_joint<T: Joint<N, M>, M>(&mut self, joint: &Rc<RefCell<T>>, activation: &mut ActivationManager<N>)

Removes a joint from this manager.

This will force the activation of the two objects attached to the joint.

fn remove(&mut self, b: &Rc<RefCell<RigidBody<N>>>, activation: &mut ActivationManager<N>)

Removes every joint attached to a given rigid body.

This will force the activation of every object attached to the deleted joints.

fn update(&mut self, activation: &mut ActivationManager<N>)

Activates the objects that interact with an activated object through a joint.

fn constraints(&mut self, constraint: &mut Vec<Constraint<N>>)

Collects all the constraints caused by joints.