Struct nphysics2d::detection::ActivationManager

pub struct ActivationManager {
    // some fields omitted
}

Structure that monitors island-based activation/deactivation of objects.

It is responsible for making objects sleep or wake up.

Methods

impl ActivationManager

fn new(mix_factor: Scalar) -> ActivationManager

Creates a new ActivationManager.

Arguments:

thresold - the minimum energy required to keep an object awake.

mix_factor - the ratio of energy to keep between two frames.

fn deferred_activate(&mut self, b: &RigidBodyHandle)

Notify the ActivationManager that is has to activate an object at the next update.

fn update(&mut self, world: &mut RigidBodyCollisionWorld, joints: &JointManager, bodies: &HashMap<usize, RigidBodyHandle, UintTWHash>)

Update the activation manager, activating and deactivating objects when needed.