Struct nphysics2d::integration::BodyForceGenerator

pub struct BodyForceGenerator<N: Scalar> {
    // some fields omitted
}

A constant linear and angular force generator.

Methods

impl<N: Scalar> BodyForceGenerator<N>

fn new(lin_acc: Vector<N>, ang_acc: Orientation<N>) -> BodyForceGenerator<N>

Creates a new BodyForceGenerator.

Arguments:

lin_acc - the linear acceleration to apply to every body on the scene.

ang_acc - the angular acceleration to apply to every body on the scene.

impl<N: Scalar> BodyForceGenerator<N>

fn lin_acc(&self) -> Vector<N>

The linear acceleration applied by this force generator.

fn set_lin_acc(&mut self, lin_acc: Vector<N>)

Sets the linear acceleration applied by this force generator.

fn ang_acc(&self) -> Orientation<N>

The angular acceleration applied by this force generator.

fn set_ang_acc(&mut self, ang_acc: Orientation<N>)

Sets the angular acceleration applied by this force generator.