Struct nphysics2d::joint::MouseConstraint

pub struct MouseConstraint<N: Real> { /* fields omitted */ }

A spring-like constraint to be used to drag a body part with the mouse.

Methods

impl<N: Real> MouseConstraint<N>

pub fn new(
b1: BodyHandle,
b2: BodyHandle,
anchor1: Point<N>,
anchor2: Point<N>,
limit: N
) -> Self

Initialize a mouse constraint between two bodies.BodyHandle

Typically, b1 will be the ground and the anchor the position of the mouse. Both anchors are expressed in the local coordinate frames of the corresponding body parts.

pub fn set_anchor_1(&mut self, anchor1: Point<N>)

Change the first anchor, expressed in the local space of the first body part.

pub fn set_anchor_2(&mut self, anchor2: Point<N>)

Change the first anchor, expressed in the local space of the second body part.