Module nphysics2d::math

Compilation flags dependent aliases for mathematical types.

Constants

ANGULAR_DIM
The maximum number of possible rotations of a rigid body.

DIM
The maximum number of possible translations of a rigid body.

SPATIAL_DIM
The maximum number of possible rotations and translations of a rigid body.

Type Definitions

AngularDim
The dimension of the rotation.

AngularInertia
The inertia tensor type.

AngularVector
The angular vector type.

Dim
The dimension of the ambiant space.

Force
The force type combining a linear force and a torque.

Inertia
The inertia type.

InertiaMatrix
The inertia matrix type.

Isometry
The transformation matrix type.

Jacobian
The type of a constraint jacobian in twist coordinates.

JacobianSlice
The type of a slice of the constraint jacobian in twist coordinates.

JacobianSliceMut
The type of a mutable slice of the constraint jacobian in twist coordinates.

Orientation
The orientation type.

Point
The point type.

Rotation
The rotation matrix type.

SpatialDim
The dimension of a spatial vector.

SpatialMatrix
Square matrix with dimension SpatialDim × SpatialDim.

SpatialVector
The vector type with dimension SpatialDim × 1.

Translation
The translation type.

Vector
The vector type.

Velocity
The velocity type combining the linear velocity and the angular velocity.