Module nphysics2d::object

Objects that may be added to the physical world.

Structs

ActivationStatus
The activation status of a body.

Bodies
Iterator yielding all the bodies on a body set.

BodyHandle
A unique identifier of a body added to the world.

BodySet
A set containing all the bodies added to the world.

ColliderData
Data stored into each collider.

Ground
A singleton representing the ground.

Material
Description of the state of surface of a solid.

Multibody
An articulated body simulated using the reduced-coordinates approach.

MultibodyLinkId
The identifier of a multibody link.

MultibodyLinkMut
Mutable reference to a multibody link.

MultibodyLinkRef
A reference to a multibody link.

MultibodyLinks
Iterator through all the multibody links.

MultibodyWorkspace
A temporary workspace for various updates of the multibody.

RigidBody
A rigid body.

Enums

Body
A body contained by the physics world.

BodyMut
A mutable body contained by the physics world.

BodyPart
A mutable body part contained by the physics world.

BodyPartMut
A mutable body part contained by the physics world.

BodyStatus
The status of a body.

Type Definitions

Collider
Type of a collider.

ColliderHandle
Type of the handle of a collider.

Colliders
Type of a reference to a collider.

Sensor
Type of a sensor.

SensorHandle
Type of the handle of a sensor.