Struct nphysics2d::object::RigidBody

pub struct RigidBody<N: Scalar> {
    // some fields omitted
}

The rigid body structure.

This is the structure describing an object on the physics world.

Methods

impl<N: Scalar> RigidBody<N>

fn position(&self) -> &Matrix<N>

Gets a reference to this body's transform.

fn shape(&self) -> &ShapeHandle<Point<N>, Matrix<N>>

A reference to this body's shape handle.

fn margin(&self) -> N

The margin surrounding this object's shape.

fn set_margin(&mut self, margin: N)

Sets the margin surrounding this object's shape.

fn center_of_mass(&self) -> &Point<N>

Gets a reference to this body's center of mass.

fn mass(&self) -> Option<N>

Gets this body's mass, if it has one.

fn restitution(&self) -> N

Gets this body's restitution coefficent.

The actual restitution coefficient of a contact is computed averaging the two bodies restitution coefficient.

fn friction(&self) -> N

Gets this body's friction coefficient.

The actual friction coefficient of a contact is computed averaging the two bodies friction coefficient.

fn is_active(&self) -> bool

Indicates whether or not this rigid body is active.

An inactive rigid body is a body that did not move for some time. It is not longer simulated by the physics engine as long as no other object touches it. It is automatically activated by the physics engine.

fn deactivation_threshold(&self) -> Option<N>

The velocity threshold bellow whith the rigid body might be deactivated.

If None, the object cannot be deactivated. If the total squared velocity (i-e: v2 + w2) falls bellow this threshold for a long enough time, the rigid body will fall asleep (i-e be "frozen") for performance reasons.

fn set_deactivation_threshold(&mut self, threshold: Option<N>)

Set the velocity threshold bellow whith the rigid body might be deactivated.

If None, the object cannot be deactivated. If the total squared velocity (i-e: v2 + w2) falls bellow this threshold for a long enough time, the rigid body will fall asleep (i-e be "frozen") for performance reasons.

fn new_dynamic<G>(shape: G, density: N, restitution: N, friction: N) -> RigidBody<N> where G: Send + Sync + Shape<Point<N>, Matrix<N>> + Volumetric<N, Point<N>, AngularInertia<N>>

Creates a new rigid body that can move.

fn new_static<G>(shape: G, restitution: N, friction: N) -> RigidBody<N> where G: Send + Sync + Shape<Point<N>, Matrix<N>>

Creates a new rigid body that cannot move.

fn new(shape: ShapeHandle<Point<N>, Matrix<N>>, mass_properties: Option<(N, Point<N>, AngularInertia<N>)>, restitution: N, friction: N) -> RigidBody<N>

Creates a new rigid body with a given shape.

Use this if the shape is shared by multiple rigid bodies. Set mass_properties to None if the rigid body is to be static.

fn can_move(&self) -> bool

Indicates whether this rigid body is static or dynamic.

fn lin_acc_scale(&self) -> Vector<N>

Gets the linear acceleraction scale of this rigid body.

fn set_lin_acc_scale(&mut self, scale: Vector<N>)

Sets the linear acceleration scale of this rigid body.

fn ang_acc_scale(&self) -> Orientation<N>

Gets the angular acceleration scale of this rigid body.

fn set_ang_acc_scale(&mut self, scale: Orientation<N>)

Sets the angular acceleration scale of this rigid body.

fn lin_vel(&self) -> Vector<N>

Get the linear velocity of this rigid body.

fn set_lin_vel(&mut self, lv: Vector<N>)

Sets the linear velocity of this rigid body.

fn lin_acc(&self) -> Vector<N>

Gets the linear acceleration of this rigid body.

fn ang_vel(&self) -> Orientation<N>

Gets the angular velocity of this rigid body.

fn set_ang_vel(&mut self, av: Orientation<N>)

Sets the angular velocity of this rigid body.

fn ang_acc(&self) -> Orientation<N>

Gets the angular acceleration of this rigid body.

fn clear_forces(&mut self)

Resets linear and angular force.

fn clear_linear_force(&mut self)

Resets linear force.

fn clear_angular_force(&mut self)

Resets angular force.

fn append_lin_force(&mut self, force: Vector<N>)

Adds an additional linear force.

fn append_ang_force(&mut self, force: Orientation<N>)

Adds an additional angular force.

fn append_force_wrt_point(&mut self, force: Vector<N>, pnt_to_com: Vector<N>)

Adds an additional force acting at a point different to the center of mass.

The pnt_to_com vector has to point from the center of mass to the point where the force acts.

fn apply_central_impulse(&mut self, impulse: Vector<N>)

Applies a one-time central impulse.

fn apply_angular_momentum(&mut self, ang_moment: Orientation<N>)

Applies a one-time angular impulse.

fn apply_impulse_wrt_point(&mut self, impulse: Vector<N>, pnt_to_com: Vector<N>)

Applies a one-time impulse to a point relative to the center of mass.

The pnt_to_com vector has to point from the center of mass to the point where the impulse acts.

fn inv_mass(&self) -> N

Gets the inverse mass of this rigid body.

fn set_inv_mass(&mut self, m: N)

Sets the inverse mass of this rigid body.

fn inv_inertia(&self) -> &AngularInertia<N>

Gets the inverse inertia tensor of this rigid body.

fn set_inv_inertia(&mut self, ii: AngularInertia<N>)

Sets the inverse inertia tensor of this rigid body.

Not that this is reset at every update by the physics engine.

fn append_transformation(&mut self, to_append: &Matrix<N>)

Appends a transformation to this rigid body.

fn prepend_transformation(&mut self, to_prepend: &Matrix<N>)

Prepends a transformation to this rigid body.

fn set_transformation(&mut self, m: Matrix<N>)

Sets the transformation of this rigid body.

fn append_translation(&mut self, t: &Vector<N>)

Appends a translation to this rigid body.

fn prepend_translation(&mut self, t: &Vector<N>)

Prepends a translation to this rigid body.

fn set_translation(&mut self, t: Vector<N>)

Stes the translation of this rigid body.

fn append_rotation(&mut self, rot: &Orientation<N>)

Appends a rotation to this rigid body.

fn prepend_rotation(&mut self, rot: &Orientation<N>)

Prepends a rotation to this rigid body.

fn set_rotation(&mut self, r: Orientation<N>)

Sets the rotation of this rigid body.

fn collision_groups(&self) -> &RigidBodyCollisionGroups

Reference to the collision groups of this rigid body.

fn set_collision_groups(&mut self, new_groups: RigidBodyCollisionGroups)

Set the new collisions groups of this rigid body.

fn user_data(&self) -> Option<&Box<Any>>

Reference to user-defined data attached to this rigid body.

fn user_data_mut(&mut self) -> Option<&mut Box<Any>>

Mutable reference to user-defined data attached to this rigid body.

fn set_user_data(&mut self, user_data: Option<Box<Any>>) -> Option<Box<Any>>

Attach some user-defined data to this rigid body and return the old one.