Enum nphysics2d::resolution::CorrectionMode

pub enum CorrectionMode {
    Velocity(Scalar),
    VelocityAndPosition(Scalar, Scalar, Scalar),
    VelocityAndPositionThresold(Scalar, Scalar, Scalar),
}

The correction coefficient used by the constraint solver.

Variants

Velocity

Penetration are solved by the penalty method.

VelocityAndPosition

Penetration are solved by the penalty method together with a hard repositioning.

VelocityAndPositionThresold

Penetration are solved by the penalty method together with a hard repositioning.

The amount of velocity correction is bounded by threshold.

Methods

impl CorrectionMode

fn vel_corr_factor(&self) -> Scalar

The velocity correction coefficient.

fn pos_corr_factor(&self) -> Scalar

The position correction coefficient.

fn min_depth_for_pos_corr(&self) -> Scalar

The minimum penetration depth required to switch on the hard repositioning based method.

fn max_depth_for_vel_corr(&self) -> Scalar

The max penetration depth the velocity correction will attempt to correct.