Struct nphysics2d::resolution::AccumulatedImpulseSolver

pub struct AccumulatedImpulseSolver<N: Scalar> {
    // some fields omitted
}

Constraint solver using the projected gauss seidel algorithm and warm-starting.

Methods

impl<N: Scalar> AccumulatedImpulseSolver<N>

fn new(step: N, correction_mode: CorrectionMode<N>, joint_corr_factor: N, rest_eps: N, num_first_order_iter: usize, num_second_order_iter: usize) -> AccumulatedImpulseSolver<N>

Creates a new AccumulatedImpulseSolver.

fn num_first_order_iter(&self) -> usize

Gets the number of iteration done by the penetration depth correction solver.

fn set_num_first_order_iter(&mut self, num: usize)

Sets the number of iteration done by the penetration depth correction solver.

fn num_second_order_iter(&self) -> usize

Gets the number of iteration done by the velocity constraint solver.

fn set_num_second_order_iter(&mut self, num: usize)

Sets the number of iteration done by the velocity constraint solver.