Function nphysics2d::solver::helper::constraints_are_ground_constraints

pub fn constraints_are_ground_constraints<N: Real>(
    body1: &BodyPart<N>, 
    body2: &BodyPart<N>
) -> bool

Test sif a constraint between the two given bodies should be a ground constraint (a constraint between a dynamic body and one without any degree of freedom).