Struct nphysics2d::solver::ImpulseCache

pub struct ImpulseCache<N> { /* fields omitted */ }

A cache for impulses.

Methods

impl<N: Copy + Zero> ImpulseCache<N>

pub fn new() -> Self

Create a new empty cache.

pub fn clear(&mut self)

Clear the cache.

pub fn contains(&self, contact_id: GenerationalId) -> bool

Test if the cache already contains the specified contact.

pub fn get(&self, contact_id: GenerationalId) -> N

Get the impulse stored for the specified contact.

Returns 0 if no cache entry is registered for this contact.

pub fn entry_id(&mut self, contact_id: GenerationalId) -> usize

Retrieve the index on the cache vector associated to the given contact.