Struct nphysics2d::detection::joint::Fixed

pub struct Fixed<N: Scalar> {
    // some fields omitted
}

A joint that prevents any relative movement (linear and angular) between two objects.

Methods

impl<N: Scalar> Fixed<N>

fn new(anchor1: Anchor<N, Matrix<N>>, anchor2: Anchor<N, Matrix<N>>) -> Fixed<N>

Creates a new Fixed joint.

fn up_to_date(&self) -> bool

Tells if the joint has been modified by the user.

fn set_local1(&mut self, local1: Matrix<N>)

Sets the the second anchor position.

The position is expressed in the second attached body’s local coordinates.

fn set_local2(&mut self, local2: Matrix<N>)

Sets the the second anchor position.

The position is expressed in the second attached body’s local coordinates.