Struct nphysics2d::detection::ActivationManager

pub struct ActivationManager<N: Real> { /* fields omitted */ }

Structure that monitors island-based activation/deactivation of bodies.

It is responsible for making objects sleep or wake up.

Methods

impl<N: Real> ActivationManager<N>

pub fn new(mix_factor: N) -> ActivationManager<N>

Creates a new ActivationManager2.

Arguments:

thresold - the minimum energy required to keep an object awake.

mix_factor - the ratio of energy to keep between two frames.

pub fn deferred_activate(&mut self, handle: BodyHandle)

Notify the ActivationManager2 that is has to activate an object at the next update.

pub fn update(
&mut self,
bodies: &mut BodySet<N>,
cworld: &CollisionWorld<N>,
constraints: &Slab<Box<JointConstraint<N>>>,
active_bodies: &mut Vec<BodyHandle>
)

Update the activation manager, activating and deactivating objects when needed.