Struct nphysics2d::world::World

pub struct World<N: Real> { /* fields omitted */ }

The physics world.

Methods

impl<N: Real> World<N>

pub fn new() -> Self

Creates a new physics world with default parameters.

The ground body is automatically created and added to the world without any colliders attached.

pub fn disable_performance_counters(&mut self)

Disable the perfomance counters that measure various times and statistics during a timestep.

pub fn enable_performance_counters(&mut self)

Enable the perfomance counters that measure various times and statistics during a timestep.

pub fn performance_counters(&self) -> &Counters

Retrieve the perfomance counters that measure various times and statistics during a timestep.

pub fn set_contact_model<C: ContactModel<N>>(&mut self, model: C)

Set the contact model for all contacts.

pub fn integration_parameters(&self) -> &IntegrationParameters<N>

Retrieve a reference to the parameters for the integration.

pub fn integration_parameters_mut(&mut self) -> &mut IntegrationParameters<N>

Retrieve a mutable reference to the parameters for the integration.

pub fn timestep(&self) -> N

Retrieve the timestep used for the integration.

pub fn set_timestep(&mut self, dt: N)

Sets the timestep used for the integration.

pub fn activate_body(&mut self, handle: BodyHandle)

Activate the given body.

pub fn add_constraint<C: JointConstraint<N>>(
&mut self,
constraint: C
) -> ConstraintHandle

Add a constraints to the physics world and retrieves its handle.

pub fn constraint(&self, handle: ConstraintHandle) -> &JointConstraint<N>

Get a reference to the specified constraint.

pub fn constraint_mut(
&mut self,
handle: ConstraintHandle
) -> &mut JointConstraint<N>

Get a mutable reference to the specified constraint.

Important traits for Box<W>
impl<W> Write for Box<W> where
W: Write + ?Sized, impl<R> Read for Box<R> where
R: Read + ?Sized, impl<I> Iterator for Box<I> where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn remove_constraint(
&mut self,
handle: ConstraintHandle
) -> Box<JointConstraint<N>>

Remove the specified constraint from the world.

pub fn add_force_generator<G: ForceGenerator<N>>(
&mut self,
force_generator: G
) -> ForceGeneratorHandle

Add a force generator to the world.

pub fn force_generator(
&self,
handle: ForceGeneratorHandle
) -> &ForceGenerator<N>

Retrieve a reference to the specified force generator.

pub fn force_generator_mut(
&mut self,
handle: ForceGeneratorHandle
) -> &mut ForceGenerator<N>

Retrieve a mutable reference to the specified force generator.

Important traits for Box<W>
impl<W> Write for Box<W> where
W: Write + ?Sized, impl<R> Read for Box<R> where
R: Read + ?Sized, impl<I> Iterator for Box<I> where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn remove_force_generator(
&mut self,
handle: ForceGeneratorHandle
) -> Box<ForceGenerator<N>>

Remove the specified force generator from the world.

pub fn set_gravity(&mut self, gravity: Vector<N>)

Set the gravity.

pub fn step(&mut self)

Execute one time step of the physics simulation.

pub fn remove_bodies(&mut self, bodies: &[BodyHandle])

Remove the specified bodies.

pub fn remove_multibody_links(&mut self, links: &[BodyHandle])

Remove several links of a single multibody.

Panics if not all links belong to the same multibody.

pub fn add_rigid_body(
&mut self,
position: Isometry<N>,
local_inertia: Inertia<N>,
local_center_of_mass: Point<N>
) -> BodyHandle

Add a rigid body to the world and retrieve its handle.

pub fn add_multibody_link<J: Joint<N>>(
&mut self,
parent: BodyHandle,
joint: J,
parent_shift: Vector<N>,
body_shift: Vector<N>,
local_inertia: Inertia<N>,
local_center_of_mass: Point<N>
) -> BodyHandle

Add a multibody link to the world and retrieve its handle.

pub fn add_collider(
&mut self,
margin: N,
shape: ShapeHandle<N>,
parent: BodyHandle,
to_parent: Isometry<N>,
material: Material<N>
) -> ColliderHandle

Add a collider to the world and retrieve its handle.

pub fn add_sensor(
&mut self,
shape: ShapeHandle<N>,
parent: BodyHandle,
to_parent: Isometry<N>
) -> SensorHandle

Add a sensor to the world and retrieve its handle.

pub fn body_part(&self, handle: BodyHandle) -> BodyPart<N>

Get a reference to the specified body part.

pub fn body(&self, handle: BodyHandle) -> Body<N>

Get a reference to the specified body.

pub fn body_mut(&mut self, handle: BodyHandle) -> BodyMut<N>

Get a mutable reference to the specified body.

pub fn multibody(&self, handle: BodyHandle) -> Option<&Multibody<N>>

Get a reference to the multibody containing the specified multibody link.

Returns None if the handle does not correspond to a multibody link in this world.

pub fn multibody_mut(&mut self, handle: BodyHandle) -> Option<&mut Multibody<N>>

Get a mutable reference to the multibody containing the specified multibody link.

Returns None if the handle does not correspond to a multibody link in this world.

pub fn multibody_link(&self, handle: BodyHandle) -> Option<MultibodyLinkRef<N>>

Get a reference to the specified multibody link.

Returns None if the handle does not correspond to a multibody link in this world.

pub fn multibody_link_mut(
&mut self,
handle: BodyHandle
) -> Option<MultibodyLinkMut<N>>

Get a mutable reference to the specified multibody link.

Returns None if the handle does not correspond to a multibody link in this world.

pub fn rigid_body(&self, handle: BodyHandle) -> Option<&RigidBody<N>>

Get a reference to the specified rigid body.

Returns None if the handle does not correspond to a rigid body in this world.

pub fn rigid_body_mut(
&mut self,
handle: BodyHandle
) -> Option<&mut RigidBody<N>>

Get a mutable reference to the specified rigid body.

Returns None if the handle does not correspond to a rigid body in this world.

pub fn collision_world(&self) -> &CollisionWorld<N>

Reference to the underlying collision world.

pub fn collision_world_mut(&mut self) -> &mut CollisionWorld<N>

Mutable reference to the underlying collision world.

pub fn collider(&self, handle: ColliderHandle) -> Option<&Collider<N>>

Get a mutable reference to the specified collider.

Returns None if the handle does not correspond to a collider in this world.

pub fn colliders(&self) -> Colliders<N>

An iterator through all the colliders on this collision world.

pub fn contact_events(&self) -> &ContactEvents

An iterator through all the contact events generated during the last execution of self.step().

pub fn proximity_events(&self) -> &ProximityEvents

An iterator through all the proximity events generated during the last execution of self.step().